[TIM-456] Removed the exposure of receptions. Only the Women Writers VRE contains a reference to receptions.
