+++ timbuctoo-rest/src/main/java/nl/knaw/huygens/timbuctoo/rest/resources/VREResource.java
@@ -22,28 +22,22 @@ package nl.knaw.huygens.timbuctoo.rest.resources;
  * #L%
  */
 
-import static nl.knaw.huygens.timbuctoo.config.Paths.SYSTEM_PREFIX;
-import static nl.knaw.huygens.timbuctoo.config.Paths.VERSION_PATH_OPTIONAL;
-
-import java.util.List;
-import java.util.Set;
+import com.google.common.collect.Sets;
+import com.google.inject.Inject;
+import nl.knaw.huygens.timbuctoo.Repository;
+import nl.knaw.huygens.timbuctoo.annotations.APIDesc;
+import nl.knaw.huygens.timbuctoo.vre.VRE;
+import nl.knaw.huygens.timbuctoo.vre.VRECollection;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
+import java.util.Set;
 
-import nl.knaw.huygens.timbuctoo.Repository;
-import nl.knaw.huygens.timbuctoo.annotations.APIDesc;
-import nl.knaw.huygens.timbuctoo.config.TypeNames;
-import nl.knaw.huygens.timbuctoo.model.RelationType;
-import nl.knaw.huygens.timbuctoo.vre.VRE;
-import nl.knaw.huygens.timbuctoo.vre.VRECollection;
-
-import com.google.common.collect.Lists;
-import com.google.common.collect.Sets;
-import com.google.inject.Inject;
+import static nl.knaw.huygens.timbuctoo.config.Paths.SYSTEM_PREFIX;
+import static nl.knaw.huygens.timbuctoo.config.Paths.VERSION_PATH_OPTIONAL;
 
 @Path(VERSION_PATH_OPTIONAL + SYSTEM_PREFIX + "/vres")
 public class VREResource extends ResourceBase {
@@ -72,74 +66,11 @@ public class VREResource extends ResourceBase {
   @Path("/{id}")
   @Produces({ MediaType.APPLICATION_JSON })
   @APIDesc("Provides info about the specified VRE.")
-  public VREInfo getVREInfo(@PathParam("id") String vreId) {
+  public VRE.VREInfo getVREInfo(@PathParam("id") String vreId) {
     VRE vre = super.getValidVRE(vreId);
 
-    VREInfo info = new VREInfo();
-    info.setName(vre.getVreId());
-    info.setDescription(vre.getDescription());
-
-    for (String name : vre.getReceptionNames()) {
-      RelationType type = repository.getRelationTypeByName(name, false);
-      if (type != null) {
-        Reception reception = new Reception();
-        reception.typeId = type.getId();
-        reception.regularName = type.getRegularName();
-        reception.inverseName = type.getInverseName();
-        reception.baseSourceType = type.getSourceTypeName();
-        reception.baseTargetType = type.getTargetTypeName();
-        reception.derivedSourceType = mapTypeName(vre, type.getSourceTypeName());
-        reception.derivedTargetType = mapTypeName(vre, type.getTargetTypeName());
-        info.addReception(reception);
-      }
-    }
+    VRE.VREInfo info = vre.toVREInfo();
     return info;
-  }
-
-  private String mapTypeName(VRE vre, String iname) {
-    return TypeNames.getInternalName(vre.mapTypeName(iname, true));
-  }
-
-  // ---------------------------------------------------------------------------
-
-  public static class VREInfo {
-    private String name;
-    private String description;
-    private final List<Reception> receptions = Lists.newArrayList();
-
-    public String getName() {
-      return name;
-    }
-
-    public void setName(String name) {
-      this.name = name;
-    }
-
-    public String getDescription() {
-      return description;
-    }
-
-    public void setDescription(String description) {
-      this.description = description;
-    }
-
-    public List<Reception> getReceptions() {
-      return receptions;
-    }
-
-    public void addReception(Reception reception) {
-      receptions.add(reception);
-    }
-  }
-
-  public static class Reception {
-    public String typeId;
-    public String regularName;
-    public String inverseName;
-    public String baseSourceType;
-    public String baseTargetType;
-    public String derivedSourceType;
-    public String derivedTargetType;
   }
 
 }
